Implement unsigned integer division and modulus.
Works just like signed integer division and modulus, except we call unsigned versions of the
runtime functions.

Extended the testmod test to test both signed and unsigned division and modulus.

Extended the test framework to report the number of tests that failed and the total number of
tests. This should make it easier to notice when tests fail.